Merge dev into main

+++ app.py
@@ -28,14 +28,12 @@ os.environ['PYTHONPYCACHEPREFIX'] = 'tmp/'
 app = Flask(__name__)
 app.secret_key = 'supersecretkey'
 
-
 # Initialize logging to output to the console
 logging.basicConfig(
     level=logging.DEBUG,  # Set logging level to DEBUG to capture all messages
     format='%(asctime)s - %(levelname)s - %(message)s',
     handlers=[logging.StreamHandler()]  # Output logs to console (stdout)
 )
-
 
 # Custom ConfigParser to preserve case sensitivity
 class CaseConfigParser(configparser.ConfigParser):
@@ -46,18 +44,14 @@ class CaseConfigParser(configparser.ConfigParser):
 config = CaseConfigParser()
 config.read('config.ini')
 
-
 # Read authentication details from config
 auth_user = config['AUTH'].get('user', 'admin')
 auth_password = config['AUTH'].get('password', 'password')
 app_port = int(config['AUTH'].get('port', '5000'))
 app_host = config['AUTH'].get('hostname', 'localhost')
 
-
 # Load the path to the JSON file from config.ini
 json_file_path = config['Paths'].get('FILTERS', 'filters.json')
-
-
 
 # Login required decorator
 def login_required(f):
@@ -67,8 +61,6 @@ def login_required(f):
             return redirect(url_for('login'))
         return f(*args, **kwargs)
     return decorated_function
-
-
 
 # Initialize database only if it doesn't exist
 def init_db():
@@ -94,21 +86,16 @@ def init_db():
         logging.info('Database already exists. No need to reload all directories.')
         return False
 
-
-
-
 #################################### CLEANUP FUNCTION #####################################################
 
 # Directory cleanup function to compare database with actual root directory
 def cleanup_orphaned_directories():
     while True:
-        datadir = config['Paths'].get('DATADIR', '').strip()
-        if not datadir:
+        data_dir = config['Paths'].get('DATADIR', '').strip()
+        if not data_dir:
             logging.error('DATADIR is not set in the configuration file.')
             return
-
         logging.info("Starting cleanup of orphaned directories...")
-
         # Connect to the SQLite database
         conn = sqlite3.connect(DIRDATABASE)
         c = conn.cursor()
@@ -118,28 +105,29 @@ def cleanup_orphaned_directories():
         db_directories = [row[0] for row in c.fetchall()]
 
         # Get all actual directories from the root directory
-        actual_directories = [d for d in os.listdir(datadir) if os.path.isdir(os.path.join(datadir, d))]
+        actual_directories = [d for d in os.listdir(data_dir) if os.path.isdir(os.path.join(data_dir, d))]
 
         # Compare the two lists and find directories in the database that don't exist in the root folder
         orphaned_directories = set(db_directories) - set(actual_directories)
+        if len(orphaned_directories) == 0:
+            logging.info(f"No orphaned directories found.")
+        else:
+            # If orphaned directories are found, delete them from the database
+            for directory in orphaned_directories:
+                logging.info(f"Removing orphaned directory: {directory} from database")
+                c.execute('DELETE FROM directories WHERE name = ?', (directory,))
 
-        # If orphaned directories are found, delete them from the database
-        for directory in orphaned_directories:
-            logging.info(f"Removing orphaned directory: {directory} from database")
-            c.execute('DELETE FROM directories WHERE name = ?', (directory,))
+                # Commit changes and close the database connection
+                conn.commit()
 
-        # Commit changes and close the database connection
-        conn.commit()
         conn.close()
-
         logging.info("Cleanup of orphaned directories completed.")
 
-        # Sleep for 10 minutes before running the check again
-        time.sleep(600)  # 600 seconds = 10 minutes
+        # Sleep for 5 minutes before running the check again
+        time.sleep(300)
 
-
-# Function to start the cleanup thread
-def start_cleanup_task():
+# Function to start the cleanup daemon
+def initiate_cleanup_daemon():
     cleanup_thread = threading.Thread(target=cleanup_orphaned_directories, daemon=True)
     cleanup_thread.start()
 
@@ -168,21 +156,21 @@ def logout():
 
 # Initialize directory data (check for new/missed directories and update their status)
 def initialize_directory_data():
-    datadir = config['Paths'].get('DATADIR', '').strip()
-    if not datadir:
+    data_dir = config['Paths'].get('DATADIR', '').strip()
+    if not data_dir:
         logging.error('DATADIR is not set in the configuration file.')
         return
 
     # Load all directories and check status (this runs once on app startup)
-    logging.debug(f'Loading all directories and checking status from: {datadir}')
-    load_directories_into_db(datadir)
+    logging.debug(f'Loading all directories and checking status from: {data_dir}')
+    load_directories_into_db(data_dir)
 
 # Load only new or modified directory data into the SQLite database
-def load_directories_into_db(datadir=None, single_directory=None):
+def load_directories_into_db(data_dir=None, single_directory=None):
     if single_directory:
         logging.debug(f'Loading single directory: {single_directory}')
-    elif datadir:
-        logging.debug(f'Starting to load directories from: {datadir}')
+    elif data_dir:
+        logging.debug(f'Starting to load directories from: {data_dir}')
     else:
         logging.error('Neither datadir nor single_directory were provided.')
         return
@@ -191,10 +179,10 @@ def load_directories_into_db(datadir=None, single_directory=None):
     if single_directory:
         directories_to_load = [single_directory]
     else:
-        if not os.path.exists(datadir):
-            logging.error(f'Data directory does not exist: {datadir}')
+        if not os.path.exists(data_dir):
+            logging.error(f'Data directory does not exist: {data_dir}')
             return
-        directories_to_load = os.scandir(datadir)
+        directories_to_load = os.scandir(data_dir)
 
     conn = sqlite3.connect(DIRDATABASE)
     c = conn.cursor()
@@ -259,8 +247,8 @@ def load_directories_into_db(datadir=None, single_directory=None):
 active_observers = {}
 
 class SubdirectoryEventHandler(FileSystemEventHandler):
-    def __init__(self, observer, dir_name):
-        self.observer = observer
+    def __init__(self, local_observer, dir_name):
+        self.observer = local_observer
         self.dir_name = dir_name
 
     def on_created(self, event):
@@ -323,33 +311,33 @@ class RootDirectoryEventHandler(FileSystemEventHandler):
             return
         
         # Create a new observer for the subdirectory
-        observer = Observer()
-        event_handler = SubdirectoryEventHandler(observer, os.path.basename(subdirectory_path))
-        observer.schedule(event_handler, path=subdirectory_path, recursive=False)
-        observer.start()
+        local_observer = Observer()
+        event_handler = SubdirectoryEventHandler(local_observer, os.path.basename(subdirectory_path))
+        local_observer.schedule(event_handler, path=subdirectory_path, recursive=False)
+        local_observer.start()
         
-        active_observers[subdirectory_path] = observer
+        active_observers[subdirectory_path] = local_observer
         logging.info(f'Started monitoring {subdirectory_path}')
 
     def stop_subdirectory_watcher(self, subdirectory_name):
         # Stop the observer for the subdirectory if it exists
-        for path, observer in list(active_observers.items()):
+        for path, local_observer in list(active_observers.items()):
             if subdirectory_name in path:
-                observer.stop()
-                observer.join()
+                local_observer.stop()
+                local_observer.join()
                 del active_observers[path]
                 logging.info(f'Stopped monitoring {subdirectory_name}')
 
 
 # Start the root directory watcher
-def start_directory_watcher(datadir):
-    logging.debug(f'Starting root directory watcher on: {datadir}')
+def start_directory_watcher(data_dir):
+    logging.debug(f'Starting root directory watcher on: {data_dir}')
     event_handler = RootDirectoryEventHandler()
-    observer = Observer()
-    observer.schedule(event_handler, path=datadir, recursive=False)
-    observer.start()
-    logging.debug(f'Root directory watcher started for: {datadir}')
-    return observer
+    local_observer = Observer()
+    local_observer.schedule(event_handler, path=data_dir, recursive=False)
+    local_observer.start()
+    logging.debug(f'Root directory watcher started for: {data_dir}')
+    return local_observer
 
 # Function to update the directory status in the database
 def update_directory_status(dir_name, new_status):
@@ -375,7 +363,6 @@ def update_directory_status(dir_name, new_status):
     conn.commit()
     conn.close()
     logging.debug(f"Status for {dir_name} updated to {new_status}")
-
 
 # Route to get directory data from the SQLite database
 @app.route('/get_directories_json')
@@ -408,14 +395,12 @@ def get_directories_json():
     # Return the JSON response with the explicit order
     return json_response
 
-
 # Home route to render the index.html template
 @app.route('/')
 @login_required
 def home():
     logging.info('Accessing the home page')
     return render_template('index.html')
-
 
 # Upload route
 @app.route('/upload', methods=['POST'])
@@ -428,7 +413,7 @@ def upload():
         return "Directory name not provided", 400  # Return 400 if directory_name is missing
 
     # Start upload in a separate thread to avoid blocking
-    upload_thread = threading.Thread(target=run_upload, args=(directory_name,))
+    upload_thread = threading.Thread(target=upload.py, args=(directory_name,))
     upload_thread.start()
 
     return "Upload started", 200
@@ -440,17 +425,17 @@ def reset_status():
     directory_name = request.form['directory_name']
     logging.info(f'Resetting status for directory: {directory_name}')
 
-    datadir = config['Paths'].get('DATADIR', '').strip()
+    data_dir = config['Paths'].get('DATADIR', '').strip()
 
-    if not datadir:
+    if not data_dir:
         logging.warning('Data Directory is not set in the configuration file.')
         flash('Data Directory is not set in the configuration file.', 'warning')
         return redirect(url_for('home'))
 
-    dir_path = os.path.join(datadir, directory_name)
+    dir_path = os.path.join(data_dir, directory_name)
 
     # Remove status directories
-    for status in ['uploading', 'uploaded', 'dupe']:
+    for status in ['uploading', 'uploaded', 'dupe', 'failed', 'processing']:
         status_dir = os.path.join(dir_path, f'.{status}')
         if os.path.isdir(status_dir):
             logging.debug(f'Removing status directory: {status_dir}')
@@ -465,14 +450,14 @@ def set_uploaded():
     directory_name = request.form['directory_name']
     logging.info(f'Setting directory as uploaded: {directory_name}')
 
-    datadir = config['Paths'].get('DATADIR', '').strip()
+    data_dir = config['Paths'].get('DATADIR', '').strip()
 
-    if not datadir:
+    if not data_dir:
         logging.warning('Data Directory is not set in the configuration file.')
         flash('Data Directory is not set in the configuration file.', 'warning')
         return redirect(url_for('home'))
 
-    dir_path = os.path.join(datadir, directory_name)
+    dir_path = os.path.join(data_dir, directory_name)
 
     # Remove any existing status folders
     for status in ['uploading', 'uploaded', 'dupe']:
@@ -492,13 +477,10 @@ def remove_status_file(dir_path, status):
         logging.debug(f'Removing status file: {status_path}')
         os.rmdir(status_path)
 
-
-
 def create_status_file(dir_path, status):
     status_path = os.path.join(dir_path, f'.{status}')
     logging.debug(f'Creating status file: {status_path}')
     os.makedirs(status_path, exist_ok=True)
-
 
 #################################### SETTINGS PAGE #####################################################
 
@@ -530,7 +512,7 @@ def settings():
         return redirect(url_for('settings'))
 
     # Load current settings from config.ini into the template
-    settings = {
+    settings_template = {
         'Header': dict(config['Header']),
         'Website': dict(config['Website']),
         'UploadForm': dict(config['UploadForm']),
@@ -542,7 +524,7 @@ def settings():
         'Settings': dict(config['Settings']),
         'MediaTools': dict(config['MediaTools'])
     }
-    return render_template('settings.html', settings=settings)
+    return render_template('settings.html', settings=settings_template)
 
 
 #################################### CATEGORY PAGE #####################################################
@@ -604,12 +586,12 @@ def edit_categories():
 @app.route('/get_status_updates', methods=['GET'])
 @login_required
 def get_status_updates():
-    datadir = config['Paths'].get('DATADIR', '').strip()
+    data_dir = config['Paths'].get('DATADIR', '').strip()
     directories = []
 
     # Loop through directories in the data directory and gather statuses
-    for dir_name in os.listdir(datadir):
-        dir_path = os.path.join(datadir, dir_name)
+    for dir_name in os.listdir(data_dir):
+        dir_path = os.path.join(data_dir, dir_name)
         if os.path.isdir(dir_path):
             status = 'none'
             if os.path.exists(os.path.join(dir_path, '.dupe')):
@@ -627,9 +609,6 @@ def get_status_updates():
 
     return jsonify(directories)
 
-
-
-
 #################################### MONITOR PAGE #####################################################
 
 # Monitor page route
@@ -637,7 +616,6 @@ def get_status_updates():
 @login_required
 def monitor():
     return render_template('monitor.html')
-
 
 @app.route('/get_terminal_output', methods=['GET'])
 @login_required
@@ -709,7 +687,7 @@ def get_logs():
     cursor = conn.cursor()
     
     cursor.execute('SELECT name, category, date, status, size, imdb_url, screenshot_url, image_url FROM uploads')
-    logs = cursor.fetchall()
+    existing_logs = cursor.fetchall()
     
     conn.close()
 
@@ -725,7 +703,7 @@ def get_logs():
             'screenshot_url': row[6],
             'image_url': row[7]
         }
-        for row in logs
+        for row in existing_logs
     ]
 
     return jsonify({'data': log_data})
@@ -748,7 +726,7 @@ if __name__ == '__main__':
         observer = start_directory_watcher(datadir)
 
     # Start the cleanup task to remove orphaned directories from the database
-    start_cleanup_task()
+    initiate_cleanup_daemon()
 
     # Path to your SSL certificate and key
     ssl_cert_path = '/etc/ssl/certs/selfsigned_cert.pem'
@@ -757,7 +735,7 @@ if __name__ == '__main__':
     try:
         # Run Flask app with SSL support
         logging.debug('Starting Flask app with SSL support...')
-        app.run(host='0.0.0.0', port=5000, ssl_context=(ssl_cert_path, ssl_key_path))
+        app.run(host='0.0.0.0', port=app_port, ssl_context=(ssl_cert_path, ssl_key_path))
     except KeyboardInterrupt:
         logging.info('Server interrupted by user.')
     finally:
@@ -766,11 +744,3 @@ if __name__ == '__main__':
             observer.stop()
             observer.join()
             logging.debug('Directory watcher stopped.')
-
-
-
-
-
-
-
-
